Added a very minimum Why Perflow page. Added link to navbar. Moved MyProject.js into appropriate folder for code structure. Edited comments in Navbar.js for minor grammar
