+++ frontend/src/App.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import './App.css';
 import { BrowserRouter as Router, Route } from "react-router-dom";
-import MyProjects from './components/MyProject';
+import MyProjects from './components/ProjectsList/MyProject';
+import WhyPerflow from './components/WhyPerflow/WhyPerflow';
 import Navbar from './components/Navbar';
 
 function App() {
@@ -11,6 +12,7 @@ function App() {
       <Router>
         <Route path="/" component={App}>
           <Route path="/MyProjects" component={MyProjects} />
+          <Route path="/WhyPerflow" component={WhyPerflow} />
         </Route>
       </Router>
     </React.Fragment>
