+++ frontend/src/components/WhyPerflow/WhyPerflow.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+class WhyPerflow extends React.Component {
+    render() {
+        return (
+            <React.Fragment>
+                <div className="text-center mt-5">
+                    <h1>Why Choose Perflow?</h1>
+                </div>
+            </React.Fragment>
+        )
+    }
+}
+
+export default WhyPerflow;
